Add Wikibase Repository

+++ mw-client/src/commonMain/kotlin/tech/antibytes/mediawiki/wikibase/model/EntityResponse.kt
@@ -12,5 +12,5 @@ import tech.antibytes.mediawiki.EntityId
 @Serializable
 internal data class EntityResponse(
     val entities: Map<EntityId, Entity> = emptyMap(),
-    val success: Short = 0
+    val success: Int = 0
 )
